Blend file of animations and control rig
Original animations were lost to a computer rebuild, but these were the likely basis of the animations for this template, in case it's useful for anybody. I believe it's using the old Mixamo control rig. Which i don't recommend, and suggest folks use Rigodotify instead, since it's focus is on compatibility across game engine retargeters